# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: ResolveAI API
  version: 1.0.0
  summary: AI-powered helpdesk and incident triage platform
  description: |
    ResolveAI triages support tickets, tracks business-hours SLAs, correlates ticket
    storms into incidents, and drafts grounded replies from a tenant knowledge base.

    **The governing principle of the whole API: AI produces signals, deterministic code
    makes decisions.** The model never returns a priority — it returns observations
    (`category`, `reportedImpact`, `paymentAffected`, `linguisticUrgency`), and a versioned
    rule set turns those into a priority. `GET /tickets/{id}/priority-rationale` exposes both
    halves separately so a wrong answer can be attributed to the model or to the policy.

    ### Conventions that apply everywhere

    - Every path is prefixed `/api/v1`. The API is versioned as a whole, never per endpoint.
    - All errors are RFC 7807 `application/problem+json`. Clients switch on `errorCode`,
      never on `detail`, which is human prose and may change.
    - All timestamps are ISO 8601, UTC, `Z`-suffixed. Never epoch millis.
    - JSON is `camelCase`; enum values are `SCREAMING_SNAKE` strings.
    - List responses are wrapped in `{ data, pagination }`, never a bare array.
    - Cursor pagination on live queues, offset pagination on slow-moving admin lists.
      Offset on a live queue is not merely slow, it is **wrong**: a ticket resolved on page 1
...
  contact:
    name: ResolveAI
...
  license:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /tickets:
    post:
      tags: [Tickets]
      operationId: createTicket
      summary: Create a ticket; triage runs asynchronously
      description: |
        Commits the ticket **and an outbox event in one transaction** and returns
        immediately. That single transaction is what removes the dual-write problem: there is
        no window in which a ticket exists but its triage job does not, and no window in
        which a job exists for a ticket that was rolled back.

        **`202`, not `201`, and this is the interview question.** The resource *is* created,
        which normally argues for `201`. But the response deliberately advertises
        `analysisStatus: "PROCESSING"`, and `priority`, `category`, `assignee` and `team` are
        all still null and **will change without any further client action**. `202` tells the
        client to poll the status resource; `201` would imply the representation is final.

        > **Phase-dependent status code.** This endpoint returns **`201` in Phase 5** and
        > **`202` from Phase 6 onward**. In Phase 5 there is no outbox and no async triage, so
        > the representation returned *is* final and `202` would be a lie. The change lands in
        > its own commit.

        **`body` is capped at 20,000 characters as a security control**, not a tidiness rule.
        Every ticket triggers an LLM call, so an uncapped body is a token-cost denial of
        service.

        **Idempotency:** replaying the same key with the same body returns the original `202`
        verbatim, including the same `id`. The same key with a *different* body is `422`,
        never a silent replay of the wrong response.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTicketRequest' }
            examples:
              customer:
                summary: Customer reporting a failed payment
                value:
                  subject: Payment deducted but order still pending
                  body: |
                    I paid via UPI at 14:03 and the money left my account, but the order page
                    still says "awaiting payment". I have the UPI reference if you need it.
                  attachmentIds: [9021]
              onBehalfOf:
                summary: Agent logging a phone call on a customer's behalf
                value:
                  subject: Cannot log in after password reset
                  body: Customer called; reset link lands but the new password is rejected.
                  onBehalfOf: 4821
      responses:
        '202':
          description: Accepted. Triage is in flight; poll `links.analysis`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TicketCreatedResponse' }
              example:
                id: 88213
                reference: TKT-10428
                subject: Payment deducted but order still pending
                status: OPEN
                priority: UNTRIAGED
                category: null
                analysisStatus: PROCESSING
                requester: { id: 4821, fullName: Priya Raman }
                assignee: null
                team: null
                createdAt: '2026-09-20T09:22:05Z'
                links:
                  self: /api/v1/tickets/88213
                  analysis: /api/v1/tickets/88213/analysis
        '400':
          description: '`VALIDATION_ERROR`, or `IDEMPOTENCY_KEY_REQUIRED` when the header is missing or malformed.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ValidationProblem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: '`FORBIDDEN` — `onBehalfOf` used by a CUSTOMER.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: '`IDEMPOTENCY_KEY_CONFLICT` — same key, different body.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: https://resolveai.dev/errors/idempotency-key-conflict
                title: Idempotency key conflict
                status: 422
                errorCode: IDEMPOTENCY_KEY_CONFLICT
                detail: This Idempotency-Key was used for a different request body.
                instance: /api/v1/tickets
                timestamp: '2026-09-20T09:22:05Z'
                traceId: 0af7651916cd43dd8448eb211c80319c
        '429': { $ref: '#/components/responses/RateLimited' }
        '503':
          description: '`IDEMPOTENCY_UNAVAILABLE` — Redis is down and idempotency fails closed.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

    get:
      tags: [Tickets]
      operationId: listTickets
      summary: List tickets — filtered, searched, cursor-paginated
      description: |
        **What the caller can see is determined by role, not by a query parameter**, and the
        scoping is applied server-side and is not overridable:

        | Role | Sees |
        |---|---|
        | `CUSTOMER` | only tickets where `requester_id = me` |
        | `AGENT` | tickets assigned to them **or** in their team's queue |
        | `TEAM_LEAD` | all tickets in their team |
        | `ADMIN` | all tickets in the tenant |

        **Cursor pagination, not `page`/`offset`, and the reason is correctness rather than
        speed.** Tickets resolve out of a filtered set while an agent is paging through it.
        With `OFFSET`, resolving a ticket on page 1 shifts everything up by one and **page 2
        silently skips a ticket the agent never saw.** The cursor encodes `(created_at, id)`
        and is stable under concurrent mutation.

        There is no `totalElements`: counting a filtered set on every page is expensive and
        nobody uses the number.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Size'
        - name: status
          in: query
          description: 'Repeatable: `?status=OPEN&status=ASSIGNED`.'
          schema:
            type: array
            items: { $ref: '#/components/schemas/TicketStatus' }
          style: form
          explode: true
        - name: priority
          in: query
          description: Repeatable.
          schema:
            type: array
            items: { $ref: '#/components/schemas/Priority' }
          style: form
          explode: true
        - name: teamId
          in: query
          description: TEAM_LEAD+ only. A CUSTOMER passing this gets `403`.
          schema: { type: integer, format: int64 }
        - name: assigneeId
          in: query
          description: Numeric id, or the literal `me`.
          schema: { type: string, examples: ['me', '1204'] }
        - name: incidentId
          in: query
          schema: { type: integer, format: int64 }
        - name: slaState
          in: query
          schema: { type: string, enum: [AT_RISK, BREACHED] }
        - name: q
          in: query
          description: Full-text over subject and body.
          schema: { type: string, maxLength: 200 }
        - name: createdFrom
          in: query
          schema: { type: string, format: date-time }
        - name: createdTo
          in: query
          schema: { type: string, format: date-time }
        - name: sort
          in: query
          schema: { type: string, enum: [created_at, priority, sla_deadline], default: created_at }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        '200':
          description: A page of ticket summaries.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CursorPage'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/TicketSummary' }
              example:
                data:
                  - id: 88213
                    reference: TKT-10428
                    subject: Payment deducted but order still pending
                    status: ASSIGNED
                    priority: P2
                    category: PAYMENT
                    requester: { id: 4821, fullName: Priya Raman }
                    assignee: { id: 1204, fullName: Arjun Mehta }
                    team: { id: 7, name: Payments }
                    incidentRef: null
                    sla:
                      firstResponse: { state: RUNNING, remainingBusinessMinutes: 142, atRisk: false }
                      resolution: { state: RUNNING, remainingBusinessMinutes: 812, atRisk: false }
                    messageCount: 2
                    createdAt: '2026-09-20T09:22:05Z'
                    updatedAt: '2026-09-20T09:22:41Z'
                pagination:
                  size: 25
                  nextCursor: eyJjIjoiMjAyNi0wOS0yMFQwOToyMjowNVoiLCJpIjo4ODIxM30
                  hasNext: true
        '400':
          description: '`INVALID_CURSOR`, or `VALIDATION_ERROR` for a bad enum or out-of-range `size`.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ValidationProblem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: '`FORBIDDEN` — a CUSTOMER passing `teamId`.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /tickets/{id}:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    get:
      tags: [Tickets]
      operationId: getTicket
      summary: Full ticket detail
      description: |
        **Field visibility by role is enforced in the response mapper by returning a
        different DTO, not by null-setting fields on a shared one.** A `CUSTOMER` never
        receives `INTERNAL` messages, `priorityRationale`, `analysisStatus`, `latestDraftId`
        or the internal SLA detail. A shared DTO with `if (role != CUSTOMER) ...` is one
        forgotten branch away from leaking internal notes; separate types make the leak a
        compile error.

        **`404`, never `403`, for a ticket in another tenant or belonging to another
        customer.** A `403` would confirm the ticket exists, which is an information leak.
        `403` is reserved for "you are authenticated, this resource is in your scope, but this
        *action* is not permitted".

        The `ETag` response header carries the entity `version` and is required back as
        `If-Match` on every mutation.
      parameters:
        - name: include
          in: query
          description: 'Comma-separated: `timeline`, `analysis`, `sla`, `incident`. Default: all except `timeline`.'
          schema: { type: string, examples: ['timeline,sla'] }
      responses:
        '200':
          description: Ticket detail.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TicketDetail' }
              example:
                id: 88213
                reference: TKT-10428
                subject: Payment deducted but order still pending
                body: I paid via UPI at 14:03 and the money left my account...
                status: ASSIGNED
                priority: P2
                category: PAYMENT
                priorityRationale:
                  policyVersion: v3
                  computed: P2
                  because: ['plan tier PRO', 'payment affected', 'impact SINGLE_USER']
                reopenCount: 0
                messages:
                  - id: 51002
                    authorId: 4821
                    authorName: Priya Raman
                    authorRole: CUSTOMER
                    visibility: PUBLIC
                    body: I paid via UPI at 14:03...
                    isFirstResponse: false
                    fromDraftId: null
                    attachments: []
                    createdAt: '2026-09-20T09:22:05Z'
                sla:
                  firstResponse:
                    state: RUNNING
                    targetBusinessMinutes: 240
                    elapsedBusinessMinutes: 98
                    remainingBusinessMinutes: 142
                    nextDeadlineAt: '2026-09-20T13:22:05Z'
                    nextRung: 50
                    escalationsFired: []
                    atRisk: false
                incident: null
                analysisStatus: READY
                latestDraftId: 9912
                _etag: W/"7"
                createdAt: '2026-09-20T09:22:05Z'
                updatedAt: '2026-09-20T11:01:12Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/TicketNotFound' }

    patch:
      tags: [Tickets]
      operationId: updateTicket
      summary: Update subject, category or team
      description: |
        `PATCH` handles plain field updates and nothing else. Anything with its own
        authorization, side effects or error vocabulary gets a sub-resource `POST`
        (`/assign`, `/status`, `/resolve`, `/reopen`) — see the deviation note in doc 05.

        TODO: schema.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Updated.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/TicketNotFound' }
        '409': { $ref: '#/components/responses/VersionConflict' }
  /tickets/{id}/messages:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    post:
      tags: [Tickets]
      operationId: addMessage
      summary: Add a public reply or an internal note
      description: |
        **An agent's first `PUBLIC` message stops the first-response SLA clock in the same
        transaction**, and `slaEffect` in the response is the important field. One transaction
        is what closes the race against the SLA poller: an agent replying at the exact instant
        a breach would fire cannot produce both a first response and a breach. Returning the
        effect means the UI does not have to re-fetch to discover it.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddMessageRequest' }
            examples:
              publicReply:
                summary: Agent reply sent from an AI draft
                value:
                  body: Thanks Priya — I can see the UPI reference and have raised a reversal with the PSP.
                  visibility: PUBLIC
                  fromDraftId: 9912
              internalNote:
                summary: Internal note, never visible to the customer
                value:
                  body: PSP dashboard shows the debit but no capture. Waiting on their ops team.
                  visibility: INTERNAL
      responses:
        '201':
          description: Message created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
              example:
                id: 51003
                ticketId: 88213
                authorId: 1204
                authorName: Arjun Mehta
                authorRole: AGENT
                visibility: PUBLIC
                body: Thanks Priya — I can see the UPI reference...
                isFirstResponse: true
                slaEffect:
                  firstResponse: MET
                  metAt: '2026-09-20T11:04:19Z'
                createdAt: '2026-09-20T11:04:19Z'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: '`FORBIDDEN` — a CUSTOMER attempting `visibility: INTERNAL`.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404': { $ref: '#/components/responses/TicketNotFound' }
        '409':
          description: '`TICKET_CLOSED` — cannot post to a CLOSED ticket.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: '`INVALID_DRAFT_REFERENCE` — `fromDraftId` is not a SHOWN draft on this ticket.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /tickets/{id}/resolve:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    post:
      tags: [Tickets]
      operationId: resolveTicket
      summary: Resolve with resolution text
      description: 'TODO: schema. Body `{ resolutionText }`. Stops the resolution clock and queues the ticket for knowledge-base indexing.'
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Resolved.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/TicketNotFound' }
        '409': { $ref: '#/components/responses/VersionConflict' }
  /tickets/{id}/reopen:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    post:
      tags: [Tickets]
      operationId: reopenTicket
      summary: Reopen a resolved or closed ticket
      description: 'TODO: schema. Body `{ reason }`. Increments `reopenCount`, which feeds the REOPEN_BUMP priority rule at >= 2.'
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Reopened.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/TicketNotFound' }
        '409': { $ref: '#/components/responses/VersionConflict' }

  # ════════════════════════════════════════════════════════════════════════════
  # TRIAGE & AI
  # ════════════════════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <accessToken>` from `POST /auth/login` or `/auth/refresh`.
        Access tokens live 900 seconds.

        **Tokens are never accepted as a query parameter.** They end up in access logs,
        browser history and `Referer` headers, none of which are under this application's
        control.
...
  responses:
    Unauthorized:
      description: '`UNAUTHORIZED` — missing, malformed or expired access token.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...